requirements.txt
bcrypt==3.1.3
cryptography==1.8.1
flask==0.12.1
flask-cors==3.0.2
flask-socketio==2.8.6
mongoengine==0.13.0
pip-save==0.2.0
pycountry==17.1.8
PyJWT==1.5.0
pymongo==3.4.0
validators==0.11.3
